woriking ish gravity

+++ src/main/java/rpg/utils/Collision.java
@@ -4,12 +4,12 @@ import rpg.game.Game;
 
 public class Collision {
 
-    public static boolean isSolid(int x, int y, int[][] lvlData) {
+    public static boolean isSolid(float x, float y, int[][] lvlData) {
         if (x < 0 || x >= Game.GAME_WIDTH) return true;
         if (y < 0 || y >= Game.GAME_HEIGHT) return true;
 
-        int xIndex = x / Game.TILE_SIZE;
-        int yIndex = y / Game.TILE_SIZE;
+        int xIndex = (int) (x / Game.TILE_SIZE);
+        int yIndex = (int) (y / Game.TILE_SIZE);
         int value = lvlData[yIndex][xIndex];
 
         return value != 11;
